+++ src/services/KakaoMapService.ts
@@ -1,5 +1,49 @@
 // src/services/KakaoMapsService.ts
-// 카카오맵 API를 전역적으로 관리하는 서비스
+// 카카오맵 API와 주소 변환 서비스를 통합 관리하는 서비스
+
+export interface DaumAddressData {
+  postcode: string;
+  postcode1: string;
+  postcode2: string;
+  postcode_seq: string;
+  zonecode: string;
+  address: string;
+  address_english: string;
+  address_type: string;
+  bcode: string;
+  bname: string;
+  bname_english: string;
+  bname1: string;
+  bname1_english: string;
+  bname2: string;
+  bname2_english: string;
+  sido: string;
+  sido_english: string;
+  sigungu: string;
+  sigungu_english: string;
+  sigungu_code: string;
+  user_language_type: string;
+  query: string;
+  building_name: string;
+  building_code: string;
+  apartment: string;
+  jibun_address: string;
+  jibun_address_english: string;
+  road_address: string;
+  road_address_english: string;
+  auto_road_address: string;
+  auto_road_address_english: string;
+  auto_jibun_address: string;
+  auto_jibun_address_english: string;
+  user_selected_type: string;
+  no_selected: string;
+  hname: string;
+  roadname_code: string;
+  roadname: string;
+  roadname_english: string;
+  latitude?: number;
+  longitude?: number;
+}
 
 // 카카오맵 API 로딩 상태를 추적하는 인터페이스
 interface KakaoMapsLoadingStatus {
@@ -154,7 +198,50 @@ export const getKakaoMapsLoadingStatus = (): KakaoMapsLoadingStatus => {
 
 // React Hook으로 사용하기 위한 export 추가 - 필요한 경우 별도로 구현할 수 있음
 
+export const convertKakaoToDaumAddress = (address: string): Promise<DaumAddressData> => {
+  return new Promise((resolve, reject) => {
+    if ((window as any).daum && (window as any).daum.Postcode) {
+      openDaumPostcode(address, resolve, reject);
+    } else {
+      const script = document.createElement('script');
+      script.src = '//t1.daumcdn.net/mapjsapi/bundle/postcode/prod/postcode.v2.js';
+      script.onload = () => openDaumPostcode(address, resolve, reject);
+      script.onerror = () => reject(new Error('다음 주소 API 로드 실패'));
+      document.head.appendChild(script);
+    }
+  });
+};
+
+const openDaumPostcode = (
+  address: string,
+  resolve: (data: DaumAddressData) => void,
+  reject: (error: Error) => void,
+) => {
+  try {
+    new (window as any).daum.Postcode({
+      oncomplete: function (data: DaumAddressData) {
+        resolve(data);
+      },
+      onclose: function (state: string) {
+        if (state === 'FORCE_CLOSE') {
+          reject(new Error('사용자가 주소 검색을 취소했습니다.'));
+        }
+      },
+      width: '100%',
+      height: '100%',
+    }).open({
+      q: address,
+      left: window.screen.width / 2 - 500 / 2,
+      top: window.screen.height / 2 - 600 / 2,
+      popupName: 'postcodePopup',
+    });
+  } catch (error) {
+    reject(error instanceof Error ? error : new Error('주소 검색 중 오류가 발생했습니다.'));
+  }
+};
+
 export default {
   initializeKakaoMaps,
   getKakaoMapsLoadingStatus,
+  convertKakaoToDaumAddress,
 };
